Replace old card lists with new versions
