+++ src/main/scala/editor/gui/TableSelectionListener.scala
@@ -1,6 +1,6 @@
 package editor.gui
 
-import _root_.editor.collection.CardList2
+import _root_.editor.collection.CardList
 import _root_.editor.gui.display.CardTable
 
 import java.awt.event.MouseEvent
@@ -18,7 +18,7 @@ import javax.swing.event.ListSelectionListener
  * 
  * @author Alec Roelke
  */
-class TableSelectionListener(frame: MainFrame, table: CardTable, list: CardList2) extends ListSelectionListener with MouseListener {
+class TableSelectionListener(frame: MainFrame, table: CardTable, list: CardList) extends ListSelectionListener with MouseListener {
   override def valueChanged(e: ListSelectionEvent) = if (!e.getValueIsAdjusting) {
     if (table.getSelectedRow >= 0)
       frame.setDisplayedCard(list(table.convertRowIndexToModel(table.getSelectedRow)).card)
